+++ handler_test.go
@@ -8,11 +8,18 @@ import (
 )
 
 func Test_handler_challenge(t *testing.T) {
+	h := &handler{
+		cfg:   &config{},
+		store: nil,
+		slack: newSlackMock(),
+	}
+
 	body := bytes.NewBufferString(`{"challenge":"challengetoken"}`)
 	req := httptest.NewRequest("POST", "/", body)
 	req.Header.Add("Content-Type", "application/json")
 	rec := httptest.NewRecorder()
-	handler(rec, req)
+
+	h.ServeHTTP(rec, req)
 
 	if rec.Code != http.StatusOK {
 		t.Errorf("status code should be 200, but %d", rec.Code)
@@ -30,12 +37,18 @@ func Test_handler_challenge(t *testing.T) {
 }
 
 func Test_handler_else(t *testing.T) {
-	cfg = &config{}
+	h := &handler{
+		cfg:   &config{},
+		store: nil,
+		slack: newSlackMock(),
+	}
+
 	body := bytes.NewBufferString(`{"event":{"text":"not number","ts":"1.23"}}`)
 	req := httptest.NewRequest("POST", "/", body)
 	req.Header.Add("Content-Type", "application/json")
 	rec := httptest.NewRecorder()
-	handler(rec, req)
+
+	h.ServeHTTP(rec, req)
 
 	if rec.Code != http.StatusNoContent {
 		t.Errorf("status code should be 204, but %d", rec.Code)
